Refactor hero highlights into cards
